pyproject.toml
[build-system]
requires = ["hatchling==1.27.0"]
build-backend = "hatchling.build"

[project]
name = "pose-filecoin-pose"
version = "0.1.0"
description = "Python-first PoSE over host memory and GPU HBM using a real Filecoin PoRep."
readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE"}
authors = [
  {name = "Repository maintainers"},
]
dependencies = [
  "PyYAML==6.0.2",
]

[project.optional-dependencies]
dev = [
  "cbor2==5.6.5",
  "grpcio==1.71.0",
  "grpcio-tools==1.71.0",
  "maturin==1.8.2",
  "protobuf==5.29.4",
  "pytest==8.3.5",
  "ruff==0.11.0",
]

[project.scripts]
pose = "pose.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/pose"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-ra"
testpaths = ["tests"]

